mvps/mvp1/requirements.txt
paramiko>=3.4.0
mcp>=1.6.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.1
